Remove Python Protos

+++ sdk/python/feast/loaders/file.py
@@ -63,13 +63,14 @@ def export_source_to_staging_location(
     # Prepare Avro file to be exported to staging location
     if isinstance(source, pd.DataFrame):
         # DataFrame provided as a source
+        uri_path = None  # type: Optional[str]
         if uri.scheme == "file":
             uri_path = uri.path
-        else:
-            uri_path = None
 
         # Remote gs staging location provided by serving
-        dir_path, file_name, source_path = export_dataframe_to_local(source, uri_path)
+        dir_path, file_name, source_path = export_dataframe_to_local(
+            df=source, dir_path=uri_path
+        )
     elif urlparse(source).scheme in ["", "file"]:
         # Local file provided as a source
         dir_path = None
@@ -108,7 +109,7 @@ def export_source_to_staging_location(
         )
 
     # Clean up, remove local staging file
-    if isinstance(source, pd.DataFrame) and len(str(dir_path)) > 4:
+    if dir_path and isinstance(source, pd.DataFrame) and len(str(dir_path)) > 4:
         shutil.rmtree(dir_path)
 
     return [staging_location_uri.rstrip("/") + "/" + file_name]
